extract SurveyResponse class from Survey
http://pm.tablexi.com/issues/80421
