+++ spec/models/external_service/survey_spec.rb
@@ -1,24 +1,7 @@
 require 'spec_helper'
 
 describe Survey do
-
-  let(:external_service_passer) { create :external_service_passer, active: true }
-  let(:external_service) { external_service_passer.external_service }
-  let(:external_service_receiver) { create :external_service_receiver, external_service: external_service }
-
-  let :params do {
-    receiver_id: external_service_receiver.receiver.id,
-    external_service_id: external_service.id,
-    external_service_passer_id: external_service_passer.id,
-    survey_url: 'http://survey.local/show',
-    survey_edit_url: 'http://survey.local/edit'
-  } end
-
-  let :deserialized_response_data do {
-    show_url: params[:survey_url],
-    edit_url: params[:survey_edit_url]
-  } end
-
+  include_context 'external service'
 
   subject(:survey) { described_class.new external_service_passer.passer, params }
 
@@ -51,32 +34,6 @@ describe Survey do
   it 'makes the external service passer inactive' do
     survey.deactivate
     expect(external_service_passer.reload).to_not be_active
-  end
-
-  it 'gives a String of URL JSON' do
-    expect(survey.response_data).to eq deserialized_response_data.to_json
-  end
-
-  it 'provides a Surveyor edit URL if an edit URL is not given' do
-    params[:survey_edit_url] = nil
-    deserialized_response_data[:edit_url] = "#{params[:survey_url]}/take"
-    expect(survey.response_data).to eq deserialized_response_data.to_json
-  end
-
-  it 'creates an external service receiver' do
-    expect(survey).to_not be_nil
-    expect { survey.save! }.to change{ ExternalServiceReceiver.count }.by 1
-    esr = ExternalServiceReceiver.last
-    expect(esr.receiver).to eq external_service_receiver.receiver
-    expect(esr.external_service).to eq external_service
-    expect(esr.response_data).to eq deserialized_response_data.to_json
-  end
-
-  it 'merges the order detail' do
-    od = external_service_receiver.receiver
-    expect(OrderDetail).to receive(:find).and_return od
-    expect(od).to receive :merge!
-    survey.save!
   end
 
 end
